Section Edit Complete

+++ app/Http/Controllers/Api/PageController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use App\Models\Information;
+use App\Models\Blog;
+use App\Models\Package;
 use App\Models\Page;
+use App\Models\Section;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
@@ -25,13 +27,53 @@ class PageController extends Controller
 
     public function sectionEditor()
     {
-        $achievement = Information::where('name', 'achievement')->first();
-        $review = Information::where('name', 'review')->first();
-        return response()->json(compact('achievement', 'review'));
+        $this->authorize('admin');
+        $header = Section::where('name', 'header')->first();
+        $activePackages = Section::where('name', 'homeSlider')->first();
+        $packages = Package::latest()->paginate(30);
+        $achievement = Section::where('name', 'achievement')->first();
+        $review = Section::where('name', 'review')->first();
+        $footer = Section::where('name', 'footer')->first();
+        return response()->json(compact('header', 'packages', 'activePackages', 'achievement', 'review', 'footer'));
+    }
+
+    public function sectionEditorPackage()
+    {
+        $this->authorize('admin');
+        $packages = Package::latest()->paginate(30);
+        return response()->json(compact('packages'));
+    }
+
+    public function updateHeader(Request $request)
+    {
+        $this->authorize('admin');
+        if (isset($request->image)) {
+            if (File::exists($request->oldImage)) {
+                unlink($request->oldImage);
+            }
+
+            $path = 'images/layouts/header/';
+            $name = $path . 'header-logo-' . time() . '.' . explode('/', explode(':', substr($request->image, 0, strpos($request->image, ';')))[1])[1];
+    
+            if (!File::exists($path)) {
+                File::makeDirectory($path, $mode = 0777, true, true);
+            }
+    
+            Image::make($request->image)->save($name);
+        } else {
+            $name = $request->oldImage;
+        }
+        $find = Section::where('name', 'header')->first();
+
+        $information = isset($find) ? $find : new Section();
+        $information->name = 'header';
+        $information->info = json_encode(['phone'=> $request->phone, 'email'=>$request->email, 'image'=>$name]);
+        $information->save();
     }
 
     public function updateAchievement(Request $request)
     {
+        $this->authorize('admin');
         foreach ($request->oldAchievementImages as $data) {
             if (File::exists($data)) {
                 unlink($data);
@@ -52,9 +94,9 @@ class PageController extends Controller
             array_push($achievements, ['image'=>$name, 'name'=>$data['name']]);
         }
 
-        $find = Information::where('name', 'achievement')->first();
+        $find = Section::where('name', 'achievement')->first();
 
-        $information = isset($find) ? $find : new Information();
+        $information = isset($find) ? $find : new Section();
         $information->name = 'achievement';
         $information->info = json_encode(['title'=> $request->title, 'subTitle'=>$request->subTitle, 'achievements'=>$achievements]);
         $information->save();
@@ -62,6 +104,7 @@ class PageController extends Controller
 
     public function updateReview(Request $request)
     {
+        $this->authorize('admin');
         foreach ($request->oldReviewImages as $data) {
             if (File::exists($data)) {
                 unlink($data);
@@ -82,11 +125,50 @@ class PageController extends Controller
             array_push($reviews, ['image'=>$name, 'name'=>$data['name'], 'message'=>$data['message']]);
         }
 
-        $find = Information::where('name', 'review')->first();
+        $find = Section::where('name', 'review')->first();
 
-        $information = isset($find) ? $find : new Information();
+        $information = isset($find) ? $find : new Section();
         $information->name = 'review';
         $information->info = json_encode(['title'=> $request->title, 'subTitle'=>$request->subTitle, 'reviews'=>$reviews]);
+        $information->save();
+    }
+
+    public function updateFooter(Request $request)
+    {
+        $this->authorize('admin');
+        $find = Section::where('name', 'footer')->first();
+        if (isset($request->image)) {
+            if (File::exists($request->oldImage)) {
+                unlink($request->oldImage);
+            }
+
+            $path = 'images/layouts/footer/';
+            $name = $path . 'footer-logo-' . time() . '.' . explode('/', explode(':', substr($request->image, 0, strpos($request->image, ';')))[1])[1];
+    
+            if (!File::exists($path)) {
+                File::makeDirectory($path, $mode = 0777, true, true);
+            }
+    
+            Image::make($request->image)->save($name);
+        } else {
+            $name = json_decode($find->info, true)['image'];
+        }
+
+
+        $information = isset($find) ? $find : new Section();
+        $information->name = 'footer';
+        $information->info = json_encode(['image'=>$name, 'message'=>$request->message, 'newsletterMessage'=> $request->newsletterMessage, 'copyright'=>$request->copyright, 'address'=> $request->address, 'social'=>$request->social, 'phone'=>$request->phone, 'email'=>$request->email]);
+        $information->save();
+    }
+
+    public function addSliderPackage(Request $request)
+    {
+        $this->authorize('admin');
+        $find = Section::where('name', 'homeSlider')->first();
+
+        $information = isset($find) ? $find : new Section();
+        $information->name = 'homeSlider';
+        $information->info = json_encode(['id'=>$request->activePackages]);
         $information->save();
     }
 }
